14.1.1-3 baseline
Former-commit-id: 851f9df229ef0c7155296922a6659a56dea8775c [formerly 94f24e0d8a0c4653a1db99d2d67fe2b029a0bc32] [formerly 851f9df229ef0c7155296922a6659a56dea8775c [formerly 94f24e0d8a0c4653a1db99d2d67fe2b029a0bc32] [formerly 5739b4f832ed8eefd3bad10f51507e51e298145c [formerly cbfd1a2e3f6b0688eb8e95ef0eaec1450b4bfc03]]]
Former-commit-id: 5739b4f832ed8eefd3bad10f51507e51e298145c
Former-commit-id: aa59400b6736c3e2ec7824e673f8a34c39b75ef5 [formerly 4a6a209aea0fee861c3c5b0754f48a71c42d56db]
Former-commit-id: 8d67ffede46a222579b7252910d9112c1a76cabd

+++ edexOsgi/com.raytheon.edex.plugin.shef/src/com/raytheon/edex/plugin/shef/database/PostTables.java
@@ -80,7 +80,6 @@ import com.raytheon.uf.edex.database.dao.DaoConfig;
  * 02/24/2012   14535      W. Kwock    Correct the duration value.
  * 11/29/2012   15530      lbousaidi   corrected posting and production time for
  * 									   latestobsvalue table.
- * 09/19/2013   16515      w. Kwock    Fix the excessive digits in rawpp,lake,height...tables
  * 
  * </pre>
  * 
@@ -1066,7 +1065,7 @@ public class PostTables {
             cs.setString(5, shefData.getExtremum().getCode());
             cs.setTimestamp(6, new Timestamp(shefData.getObservationTimeObj()
                     .getTime()));
-            cs.setDouble(7, Double.parseDouble(dataValue));
+            cs.setFloat(7, Float.parseFloat(dataValue));
             cs.setString(8, qualifier);
             cs.setInt(9, (int) qualityCode);
 
@@ -1185,7 +1184,7 @@ public class PostTables {
             cs.setString(5, shefData.getExtremum().getCode());
             cs.setTimestamp(6, new java.sql.Timestamp(shefData
                     .getObservationTimeObj().getTime()));
-            cs.setDouble(7, Double.parseDouble(dataValue));
+            cs.setFloat(7, Float.parseFloat(dataValue));
             cs.setString(8, qualifier);
             cs.setInt(9, (int) qualityCode);
 
@@ -1317,7 +1316,7 @@ public class PostTables {
             timeStamp = new java.sql.Timestamp(basisDate.getTime());
             cs.setTimestamp(8, timeStamp);
 
-            cs.setDouble(9, Double.parseDouble(dataValue));
+            cs.setFloat(9, Float.parseFloat(dataValue));
 
             cs.setString(10, qualifier);
 
@@ -1535,7 +1534,7 @@ public class PostTables {
             ps.setTimestamp(8, timeStamp2);
 
             // ps.setFloat(9, Float.parseFloat(shefDataValue.getStringValue()));
-            ps.setDouble(9, shefDataValue.getValue().floatValue());
+            ps.setFloat(9, shefDataValue.getValue().floatValue());
 
             if (updateFlag) {
                 ps.setString(10, lid);
